Authorizations (#3)

+++ Models/Solution.cs
@@ -2,8 +2,8 @@ namespace rowi_practice.Models;
 
 public class Solution 
 {
-    public long Id { get; set; }
-    public long Author_Id { get; set; }
+    public int Id { get; set; }
+    public int Author_Id { get; set; }
     public string Reference { get; set; } 
-    public long Problem_id {get; set; }
+    public int Problem_id {get; set; }
 }
